[MIG][event_contact] Migration to v10
Document new behavior for event type contacts, and fix its test.
