+++ event_contact/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-# © 2016 OpenSynergy Indonesia
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import models
