+++ event_contact/__manifest__.py
@@ -1,13 +1,14 @@
 # -*- coding: utf-8 -*-
-# © 2016 OpenSynergy Indonesia
+# Copyright 2016 OpenSynergy Indonesia
+# Copyright 2016-2017 Jairo Llopis <jairo.llopis@tecnativa.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 {
     'name': 'Event Contacts',
-    'version': '8.0.1.0.0',
+    'version': '10.0.1.0.0',
     'summary': 'Add contacts to event and event type',
     'author': 'OpenSynergy Indonesia, '
-              'Antiun Ingeniería S.L., '
+              'Tecnativa, '
               'Odoo Community Association (OCA)',
     'website': 'https://opensynergy-indonesia.com',
     'category': 'Marketing',
