+++ event_contact/views/event_event_view.xml
@@ -1,17 +1,15 @@
 <?xml version="1.0"?>
-<openerp>
-    <data>
+<odoo>
 
-        <record model="ir.ui.view" id="event_contacts_form_view">
-            <field name="name">Event Contacts</field>
-            <field name="model">event.event</field>
-            <field name="inherit_id" ref="event.view_event_form"/>
-            <field name="arch" type="xml">
-                <field name="organizer_id" position="after">
-                    <field name="contact_ids" widget="many2many_tags"/>
-                </field>
+    <record model="ir.ui.view" id="event_contacts_form_view">
+        <field name="name">Event Contacts</field>
+        <field name="model">event.event</field>
+        <field name="inherit_id" ref="event.view_event_form"/>
+        <field name="arch" type="xml">
+            <field name="organizer_id" position="after">
+                <field name="contact_ids" widget="many2many_tags"/>
             </field>
-        </record>
+        </field>
+    </record>
 
-    </data>
-</openerp>
+</odoo>
