+++ event_contact/views/event_type_view.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- © 2016 Antiun Ingeniería S.L. - Jairo Llopis
+<!-- Copyright 2016-2017 Jairo Llopis <jairo.llopis@tecnativa.com>
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
 
-<openerp>
-<data>
+<odoo>
 
 <record id="view_event_type_form" model="ir.ui.view">
     <field name="name">Add contacts</field>
@@ -16,5 +15,4 @@
     </field>
 </record>
 
-</data>
-</openerp>
+</odoo>
